backend/requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.16,<4.0
django-cors-headers>=4.7,<5.0
PyMySQL>=1.1,<2.0
python-dotenv>=1.0,<2.0
gunicorn>=21.2,<22.0
whitenoise>=6.5,<7.0
dj-database-url>=2.0,<3.0
